style(ws): move function
