+++ backend/matcha/db/notification.py
@@ -1,6 +1,4 @@
-from flask_socketio import emit
-
-from matcha.websocket.socket_manager import SocketManager
+from matcha.websocket.notification import ws_send_notification
 
 from matcha.db.utils import (
     db_query,
@@ -32,18 +30,6 @@ def db_get_notification(id_user):
         )
 
     return array
-
-
-def ws_send_notification(id_user: int, title: str, content: str):
-    notification_message = {
-        "content": content,
-        "title": title,
-    }
-
-    sid = SocketManager().get_sid(id_user)
-
-    if sid is not None:
-        emit(title, notification_message, to=sid, namespace="/")
 
 
 def db_put_notification(id_user, title: str, content: str):
